requirements.txt
Django
psycopg2
djangorestframework
djoser
drf-yasg
django-filter
social-auth-app-django

